backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
tensorflow-cpu==2.15.0
librosa==0.10.1
numpy==1.26.3
joblib==1.3.2
soundfile==0.12.1
